app/cogs/mod-disc-bot-features/commands/where/requirements.txt
Pillow===9.4.0
aiohttp===3.8.4